+++ docs/attribute_curator/CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this gear are documented in this file.
 ## 1.3.0
 
 * Updates `nacc-attribute-deriver` to `2.1.2` which mainly changes where the deriver gets the MEDS visitdate
-* Updates to drop and report error when two forms in the same scope have the same visitdate
+* Updates to report an error when two forms in the same scope have the same visitdate; fails the whole subject
 * Adds `_uds_visitdate` to table for scopes in the same session as an UDS visit
-* Refactors scheduling models to separate raw data model dumping and post-computation for dates/scopes/etc.
+* Refactors scheduling models
 
 ## 1.2.1